Improved gyro sending speed

+++ embedded/ESP32-Display-Board/include/utilities.h
@@ -4,9 +4,9 @@
 #include <Arduino.h>
 
 struct GyroData {
-    float x;
-    float y;
-    // float z; Z is redundant for moving a cursor on a 2d surface
+    int x;
+    int y;
+    // int z; Z is redundant for moving a cursor on a 2d surface
 };
 
 #define BUTTON_PIN 22
